was changed after comments

+++ src/com/company/lesson1/table_creating.sql
@@ -1,13 +1,13 @@
 CREATE TABLE publisher
 (
-    id INT PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     org_name VARCHAR(128) NOT NULL ,
     address TEXT NOT NULL
 );
 
 CREATE TABLE book
 (
-    id INT PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     title TEXT NOT NULL,
     isbn VARCHAR(32) NOT NULL
 );
